chore: move asset loading to ryot sub-crates
