+++ apps/ryot_compass/src/lmdb.rs
@@ -7,7 +7,7 @@ use ryot::bevy_ryot::lmdb::{
     compact_map, read_area, reload_visible_area, LmdbCompactor, LmdbEnv,
     LmdbPlugin as RyotLmdbPlugin,
 };
-use ryot::bevy_ryot::{GameObjectBundle, InternalContentState};
+use ryot::bevy_ryot::GameObjectBundle;
 use ryot::prelude::drawing::TileComponent;
 use ryot::prelude::*;
 use std::collections::HashMap;
